feat(vision): continue vision work
Moves vision related functionality into static functions in a separate Vision class. This is intended to make the code easier to track. The actual commands will still be in drive, calling static functions in the vision class.

This also implements more of the branch alignment code. Still more work to be done to implement the alignment command and probably a bunch of fixes to the implementation once we can test.